Add Cloud function with GCS trigger
Source code for function is stored in separate bucket
Location had to be forced to europe-west1 has gen1 functions
do not support europe-west4 location.
gen1 function was chosen due to many errors that could not be explained
in a gen2 function. This should be investigated as soon as there is
time.